pyproject.toml
[project]
name = "langgraph-assistent-with-mcp"
version = "0.0.1"
description = "LangGraph solution template for MCP"
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "requests>=2.32.4",
    "python-dotenv>=1.1.1",
    "mcp>=1.12.0",
    "langchain>=0.3,<0.4",
    "langchain-core>=0.3,<0.4",
    "langchain-openai>=0.3.28",
    "langchain-mcp-adapters>=0.1.9",
    "langchain_community>=0.3.27",
    "langgraph>=0.5.3",
    "langgraph-cli[inmem]>=0.3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.1",
    "black>=25.1.0",
    "flake8>=7.3.0",
    "mypy>=1.17.0",
    "faiss-cpu>=1.11.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"